refactor(q8): move schedule telemetry helpers
